Length bin should not be greater than 100cm
The last length bin is 99cm+. However, prior to this fix the length bin
for some fish was 100. This caused the calculation of `biomass_vulnerable`
to be erroneous because the some fish would "walk out" of the
`selectivity_at_length` array. This was only apparent for the REC methof
because it is at the end of that array.